📝♻ Add segmented-control example modules
