+++ apps/cookbook/src/app/examples/segmented-control-example/_segmented-control-example.shared.scss
@@ -1,0 +1,15 @@
+@import './libs/designsystem/src/lib/scss/utils';
+
+:host {
+  display: block;
+  margin-bottom: size('s');
+}
+
+kirby-chip,
+kirby-segmented-control {
+  margin-bottom: size('l');
+}
+
+kirby-card {
+  margin-bottom: size('s');
+}
