+++ apps/cookbook/src/app/examples/segmented-control-example/segmented-control-example.component.html
@@ -1,0 +1,1 @@
+<cookbook-segmented-control-example-default></cookbook-segmented-control-example-default>
