+++ apps/cookbook/src/app/examples/segmented-control-example/segmented-control-example.component.scss
@@ -9,20 +9,17 @@
   @include media('>=medium') {
     padding: size('l');
   }
-}
 
-kirby-segmented-control {
-  margin-bottom: size('m');
-}
+  h3 {
+    font-size: font-size('xs');
+    color: get-color('semi-dark');
+    text-transform: uppercase;
+    margin-top: size('xs');
+    margin-bottom: size('xxxs');
+  }
 
-.segment-data {
-  padding: size('xxl');
-  text-align: center;
-  background: get-color('white');
-}
-
-fieldset {
-  display: inline-block;
-  margin-top: size('l');
-  margin-bottom: size('s');
+  > * {
+    display: block;
+    margin: size('xxs') 0;
+  }
 }
