+++ apps/cookbook/src/app/showcase/segmented-control-showcase/segmented-control-showcase.component.html
@@ -1,8 +1,9 @@
 <div class="example">
+  <h3>Default</h3>
   <cookbook-example-viewer [html]="example.template">
-    <div>
-      <cookbook-segmented-control-example #example></cookbook-segmented-control-example>
-    </div>
+    <cookbook-segmented-control-example-default
+      #example
+    ></cookbook-segmented-control-example-default>
   </cookbook-example-viewer>
 
   <h2>Properties:</h2>
